% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/election.R
\name{election}
\alias{election}
\title{Election results 2014}
\usage{
election(election, type)
}
\arguments{
\item{election}{Type of election. Must be one of the following: "Riksdagsval", "Landstingsval" or "Kommunval"}

\item{type}{Specifies which level the results should be presented in. Must be either "Kommun" or "Valdistrikt"}
}
\value{
A list containing the name of the file and a data frame containing the election results depending 
on the arguments election and type
}
\description{
A function to obtain election results from the Swedish election 2014
}
