% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file.R
\name{get_file}
\alias{get_file}
\title{Obtain election files}
\usage{
get_file(x)
}
\arguments{
\item{x}{Specifying what election type to obtain with three allowed arguments ("Riksdagsval", "Landstingsval" or "Kommunval")}
}
\value{
A list containing the election type and the election result on municipality level
}
\description{
A function to obtain election files from the Swedish election 2014
}
