% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_election.R
\name{graph_election}
\alias{graph_election}
\title{Plot the election results}
\usage{
graph_election(municipality)
}
\arguments{
\item{municipality}{The name of the municipality to plot}
}
\value{
A bar chart over the results in chosen municipality
}
\description{
A function to plot the results from the Swedish election 2014 for a specific municipality in Sweden. 
Only the nine biggest parties in Sweden are plotted
}
