\name{lab5sarasara-package}
\alias{lab5sarasara-package}
\alias{lab5sarasara}
\docType{package}
\title{
\packageTitle{lab5sarasara}
}
\description{
\packageDescription{lab5sarasara}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lab5sarasara}
\packageIndices{lab5sarasara}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{lab5sarasara}

Maintainer: \packageMaintainer{lab5sarasara}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
