% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voters.R
\name{voters}
\alias{voters}
\title{Voters 2014}
\usage{
voters(election)
}
\arguments{
\item{election}{Type of election. Must be one of the following: "Riksdagsval", "Landstingsval" or "Kommunval"}
}
\value{
A list containing the name of the file and a data frame containing information about the voters from the election
}
\description{
A function to obtain information about the voters from the Swedish election 2014
}
