% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file.R
\name{get_file}
\alias{get_file}
\title{Obtain election files}
\usage{
get_file(x)
}
\arguments{
\item{x}{Number specifying which file to get where x is between 1 and 18 and of length one}
}
\value{
A list containing the file name and a data frame of the file
}
\description{
A function to obtain election files from the Swedish election 2014
}
