% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file.R
\name{get_file}
\alias{get_file}
\title{Obtain election files}
\usage{
get_file()
}
\value{
A data frame with the election result on municipality level
}
\description{
A function to obtain election files from the Swedish parliamentary election 2014
}
