% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab5sarasara.R
\docType{package}
\name{lab5sarasara}
\alias{lab5sarasara}
\alias{lab5sarasara-package}
\title{Election results of the Swedish election 2014}
\description{
A package containing functions to investigate the election results of the Swedish election 2014
}
\author{
Sara Jesperson and Sara Johansson
}
